Struct rusoto_mediaconvert::H265Settings

pub struct H265Settings {
    pub adaptive_quantization: Option<String>,
    pub alternate_transfer_function_sei: Option<String>,
    pub bitrate: Option<i64>,
    pub codec_level: Option<String>,
    pub codec_profile: Option<String>,
    pub dynamic_sub_gop: Option<String>,
    pub flicker_adaptive_quantization: Option<String>,
    pub framerate_control: Option<String>,
    pub framerate_conversion_algorithm: Option<String>,
    pub framerate_denominator: Option<i64>,
    pub framerate_numerator: Option<i64>,
    pub gop_b_reference: Option<String>,
    pub gop_closed_cadence: Option<i64>,
    pub gop_size: Option<f64>,
    pub gop_size_units: Option<String>,
    pub hrd_buffer_initial_fill_percentage: Option<i64>,
    pub hrd_buffer_size: Option<i64>,
    pub interlace_mode: Option<String>,
    pub max_bitrate: Option<i64>,
    pub min_i_interval: Option<i64>,
    pub number_b_frames_between_reference_frames: Option<i64>,
    pub number_reference_frames: Option<i64>,
    pub par_control: Option<String>,
    pub par_denominator: Option<i64>,
    pub par_numerator: Option<i64>,
    pub quality_tuning_level: Option<String>,
    pub qvbr_settings: Option<H265QvbrSettings>,
    pub rate_control_mode: Option<String>,
    pub sample_adaptive_offset_filter_mode: Option<String>,
    pub scene_change_detect: Option<String>,
    pub slices: Option<i64>,
    pub slow_pal: Option<String>,
    pub spatial_adaptive_quantization: Option<String>,
    pub telecine: Option<String>,
    pub temporal_adaptive_quantization: Option<String>,
    pub temporal_ids: Option<String>,
    pub tiles: Option<String>,
    pub unregistered_sei_timecode: Option<String>,
    pub write_mp_4_packaging_type: Option<String>,
}

Settings for H265 codec

Fields
adaptive_quantization: Option<String>

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

alternate_transfer_function_sei: Option<String>

Enables Alternate Transfer Function SEI message for outputs using Hybrid Log Gamma (HLG) Electro-Optical Transfer Function (EOTF).

bitrate: Option<i64>

Average bitrate in bits/second. Required for VBR and CBR. For MS Smooth outputs, bitrates must be unique when rounded down to the nearest multiple of 1000.

codec_level: Option<String>

H.265 Level.

codec_profile: Option<String>

Represents the Profile and Tier, per the HEVC (H.265) specification. Selections are grouped as [Profile] / [Tier], so "Main/High" represents Main Profile with High Tier. 4:2:2 profiles are only available with the HEVC 4:2:2 License.

dynamic_sub_gop: Option<String>

Choose Adaptive to improve subjective video quality for high-motion content. This will cause the service to use fewer B-frames (which infer information based on other frames) for high-motion portions of the video and more B-frames for low-motion portions. The maximum number of B-frames is limited by the value you provide for the setting B frames between reference frames (numberBFramesBetweenReferenceFrames).

flicker_adaptive_quantization: Option<String>

Adjust quantization within each frame to reduce flicker or 'pop' on I-frames.

framerate_control: Option<String>

If you are using the console, use the Framerate setting to specify the frame rate for this output. If you want to keep the same frame rate as the input video, choose Follow source. If you want to do frame rate conversion, choose a frame rate from the dropdown list or choose Custom. The framerates shown in the dropdown list are decimal approximations of fractions. If you choose Custom, specify your frame rate as a fraction. If you are creating your transcoding job sepecification as a JSON file without the console, use FramerateControl to specify which value the service uses for the frame rate for this output. Choose INITIALIZEFROMSOURCE if you want the service to use the frame rate from the input. Choose SPECIFIED if you want the service to use the frame rate you specify in the settings FramerateNumerator and FramerateDenominator.

framerate_conversion_algorithm: Option<String>

When set to INTERPOLATE, produces smoother motion during frame rate conversion.

framerate_denominator: Option<i64>

Frame rate denominator.

framerate_numerator: Option<i64>

Frame rate numerator - frame rate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

gop_b_reference: Option<String>

If enable, use reference B frames for GOP structures that have B frames > 1.

gop_closed_cadence: Option<i64>

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

gop_size: Option<f64>

GOP Length (keyframe interval) in frames or seconds. Must be greater than zero.

gop_size_units: Option<String>

Indicates if the GOP Size in H265 is specified in frames or seconds. If seconds the system will convert the GOP Size into a frame count at run time.

hrd_buffer_initial_fill_percentage: Option<i64>

Percentage of the buffer that should initially be filled (HRD buffer model).

hrd_buffer_size: Option<i64>

Size of buffer (HRD buffer model) in bits. For example, enter five megabits as 5000000.

interlace_mode: Option<String>

Use Interlace mode (InterlaceMode) to choose the scan line type for the output. * Top Field First (TOPFIELD) and Bottom Field First (BOTTOMFIELD) produce interlaced output with the entire output having the same field polarity (top or bottom first). * Follow, Default Top (FOLLOWTOPFIELD) and Follow, Default Bottom (FOLLOWBOTTOMFIELD) use the same field polarity as the source. Therefore, behavior depends on the input scan type. - If the source is interlaced, the output will be interlaced with the same polarity as the source (it will follow the source). The output could therefore be a mix of "top field first" and "bottom field first". - If the source is progressive, the output will be interlaced with "top field first" or "bottom field first" polarity, depending on which of the Follow options you chose.

max_bitrate: Option<i64>

Maximum bitrate in bits/second. For example, enter five megabits per second as 5000000. Required when Rate control mode is QVBR.

min_i_interval: Option<i64>

Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. This setting is only used when Scene Change Detect is enabled. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

number_b_frames_between_reference_frames: Option<i64>

Number of B-frames between reference frames.

number_reference_frames: Option<i64>

Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

par_control: Option<String>

Using the API, enable ParFollowSource if you want the service to use the pixel aspect ratio from the input. Using the console, do this by choosing Follow source for Pixel aspect ratio.

par_denominator: Option<i64>

Pixel Aspect Ratio denominator.

par_numerator: Option<i64>

Pixel Aspect Ratio numerator.

quality_tuning_level: Option<String>

Use Quality tuning level (H265QualityTuningLevel) to specifiy whether to use fast single-pass, high-quality singlepass, or high-quality multipass video encoding.

qvbr_settings: Option<H265QvbrSettings>

Settings for quality-defined variable bitrate encoding with the H.265 codec. Required when you set Rate control mode to QVBR. Not valid when you set Rate control mode to a value other than QVBR, or when you don't define Rate control mode.

rate_control_mode: Option<String>

Use this setting to specify whether this output has a variable bitrate (VBR), constant bitrate (CBR) or quality-defined variable bitrate (QVBR).

sample_adaptive_offset_filter_mode: Option<String>

Specify Sample Adaptive Offset (SAO) filter strength. Adaptive mode dynamically selects best strength based on content

scene_change_detect: Option<String>

Scene change detection (inserts I-frames on scene changes).

slices: Option<i64>

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures.

slow_pal: Option<String>

Enables Slow PAL rate conversion. 23.976fps and 24fps input is relabeled as 25fps, and audio is sped up correspondingly.

spatial_adaptive_quantization: Option<String>

Adjust quantization within each frame based on spatial variation of content complexity.

telecine: Option<String>

This field applies only if the Streams > Advanced > Framerate (framerate) field is set to 29.970. This field works with the Streams > Advanced > Preprocessors > Deinterlacer field (deinterlacemode) and the Streams > Advanced > Interlaced Mode field (interlacemode) to identify the scan type for the output: Progressive, Interlaced, Hard Telecine or Soft Telecine. - Hard: produces 29.97i output from 23.976 input. - Soft: produces 23.976; the player converts this output to 29.97i.

temporal_adaptive_quantization: Option<String>

Adjust quantization within each frame based on temporal variation of content complexity.

temporal_ids: Option<String>

Enables temporal layer identifiers in the encoded bitstream. Up to 3 layers are supported depending on GOP structure: I- and P-frames form one layer, reference B-frames can form a second layer and non-reference b-frames can form a third layer. Decoders can optionally decode only the lower temporal layers to generate a lower frame rate output. For example, given a bitstream with temporal IDs and with b-frames = 1 (i.e. IbPbPb display order), a decoder could decode all the frames for full frame rate output or only the I and P frames (lowest temporal layer) for a half frame rate output.

tiles: Option<String>

Enable use of tiles, allowing horizontal as well as vertical subdivision of the encoded pictures.

unregistered_sei_timecode: Option<String>

Inserts timecode for each frame as 4 bytes of an unregistered SEI message.

write_mp_4_packaging_type: Option<String>

Use this setting only for outputs encoded with H.265 that are in CMAF or DASH output groups. If you include writeMp4PackagingType in your JSON job specification for other outputs, your video might not work properly with downstream systems and video players. If the location of parameter set NAL units don't matter in your workflow, ignore this setting. The service defaults to marking your output as HEV1. Choose HVC1 to mark your output as HVC1. This makes your output compliant with this specification: ISO IECJTC1 SC29 N13798 Text ISO/IEC FDIS 14496-15 3rd Edition. For these outputs, the service stores parameter set NAL units in the sample headers but not in the samples directly. Keep the default HEV1 to mark your output as HEV1. For these outputs, the service writes parameter set NAL units directly into the samples.