Struct rusoto_mediaconvert::M2tsSettings

pub struct M2tsSettings {
    pub audio_buffer_model: Option<String>,
    pub audio_frames_per_pes: Option<i64>,
    pub audio_pids: Option<Vec<i64>>,
    pub bitrate: Option<i64>,
    pub buffer_model: Option<String>,
    pub dvb_nit_settings: Option<DvbNitSettings>,
    pub dvb_sdt_settings: Option<DvbSdtSettings>,
    pub dvb_sub_pids: Option<Vec<i64>>,
    pub dvb_tdt_settings: Option<DvbTdtSettings>,
    pub dvb_teletext_pid: Option<i64>,
    pub ebp_audio_interval: Option<String>,
    pub ebp_placement: Option<String>,
    pub es_rate_in_pes: Option<String>,
    pub force_ts_video_ebp_order: Option<String>,
    pub fragment_time: Option<f64>,
    pub max_pcr_interval: Option<i64>,
    pub min_ebp_interval: Option<i64>,
    pub nielsen_id_3: Option<String>,
    pub null_packet_bitrate: Option<f64>,
    pub pat_interval: Option<i64>,
    pub pcr_control: Option<String>,
    pub pcr_pid: Option<i64>,
    pub pmt_interval: Option<i64>,
    pub pmt_pid: Option<i64>,
    pub private_metadata_pid: Option<i64>,
    pub program_number: Option<i64>,
    pub rate_mode: Option<String>,
    pub scte_35_esam: Option<M2tsScte35Esam>,
    pub scte_35_pid: Option<i64>,
    pub scte_35_source: Option<String>,
    pub segmentation_markers: Option<String>,
    pub segmentation_style: Option<String>,
    pub segmentation_time: Option<f64>,
    pub timed_metadata_pid: Option<i64>,
    pub transport_stream_id: Option<i64>,
    pub video_pid: Option<i64>,
}

MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

Fields
audio_buffer_model: Option<String>

Selects between the DVB and ATSC buffer models for Dolby Digital audio.

audio_frames_per_pes: Option<i64>

The number of audio frames to insert for each PES packet.

audio_pids: Option<Vec<i64>>

Specify the packet identifiers (PIDs) for any elementary audio streams you include in this output. Specify multiple PIDs as a JSON array. Default is the range 482-492.

bitrate: Option<i64>

Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

buffer_model: Option<String>

Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

dvb_nit_settings: Option<DvbNitSettings>

Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

dvb_sdt_settings: Option<DvbSdtSettings>

Inserts DVB Service Description Table (NIT) at the specified table repetition interval.

dvb_sub_pids: Option<Vec<i64>>

Specify the packet identifiers (PIDs) for DVB subtitle data included in this output. Specify multiple PIDs as a JSON array. Default is the range 460-479.

dvb_tdt_settings: Option<DvbTdtSettings>

Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

dvb_teletext_pid: Option<i64>

Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

ebp_audio_interval: Option<String>

When set to VIDEOANDFIXEDINTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEOINTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

ebp_placement: Option<String>

Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

es_rate_in_pes: Option<String>

Controls whether to include the ES Rate field in the PES header.

force_ts_video_ebp_order: Option<String>

Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

fragment_time: Option<f64>

The length, in seconds, of each fragment. Only used with EBP markers.

max_pcr_interval: Option<i64>

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

min_ebp_interval: Option<i64>

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is "stretched" to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

nielsen_id_3: Option<String>

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

null_packet_bitrate: Option<f64>

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

pat_interval: Option<i64>

The number of milliseconds between instances of this table in the output transport stream.

pcr_control: Option<String>

When set to PCREVERYPES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

pcr_pid: Option<i64>

Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

pmt_interval: Option<i64>

Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

pmt_pid: Option<i64>

Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

private_metadata_pid: Option<i64>

Specify the packet identifier (PID) of the private metadata stream. Default is 503.

program_number: Option<i64>

Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

rate_mode: Option<String>

When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

scte_35_esam: Option<M2tsScte35Esam>

Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

scte_35_pid: Option<i64>

Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

scte_35_source: Option<String>

Enables SCTE-35 passthrough (scte35Source) to pass any SCTE-35 signals from input to output.

segmentation_markers: Option<String>

Inserts segmentation markers at each segmentationtime period. raisegstart sets the Random Access Indicator bit in the adaptation field. raiadapt sets the RAI bit and adds the current timecode in the private data bytes. psisegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

segmentation_style: Option<String>

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of "resetcadence" is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentationtime seconds. When a segmentation style of "maintaincadence" is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationtime seconds. Note that EBP lookahead is a slight exception to this rule.

segmentation_time: Option<f64>

Specify the length, in seconds, of each segment. Required unless markers is set to none.

timed_metadata_pid: Option<i64>

Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

transport_stream_id: Option<i64>

Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

video_pid: Option<i64>

Specify the packet identifier (PID) of the elementary video stream in the transport stream.