Struct rusoto_mediaconvert::MediaConvertClient

pub struct MediaConvertClient { /* fields omitted */ }

A client for the MediaConvert API.

Methods

impl MediaConvertClient

pub fn new(region: Region) -> MediaConvertClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MediaConvertClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,