Struct rusoto_mediaconvert::Mpeg2Settings

pub struct Mpeg2Settings {
    pub adaptive_quantization: Option<String>,
    pub bitrate: Option<i64>,
    pub codec_level: Option<String>,
    pub codec_profile: Option<String>,
    pub dynamic_sub_gop: Option<String>,
    pub framerate_control: Option<String>,
    pub framerate_conversion_algorithm: Option<String>,
    pub framerate_denominator: Option<i64>,
    pub framerate_numerator: Option<i64>,
    pub gop_closed_cadence: Option<i64>,
    pub gop_size: Option<f64>,
    pub gop_size_units: Option<String>,
    pub hrd_buffer_initial_fill_percentage: Option<i64>,
    pub hrd_buffer_size: Option<i64>,
    pub interlace_mode: Option<String>,
    pub intra_dc_precision: Option<String>,
    pub max_bitrate: Option<i64>,
    pub min_i_interval: Option<i64>,
    pub number_b_frames_between_reference_frames: Option<i64>,
    pub par_control: Option<String>,
    pub par_denominator: Option<i64>,
    pub par_numerator: Option<i64>,
    pub quality_tuning_level: Option<String>,
    pub rate_control_mode: Option<String>,
    pub scene_change_detect: Option<String>,
    pub slow_pal: Option<String>,
    pub softness: Option<i64>,
    pub spatial_adaptive_quantization: Option<String>,
    pub syntax: Option<String>,
    pub telecine: Option<String>,
    pub temporal_adaptive_quantization: Option<String>,
}

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value MPEG2.

Fields
adaptive_quantization: Option<String>

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

bitrate: Option<i64>

Average bitrate in bits/second. Required for VBR and CBR. For MS Smooth outputs, bitrates must be unique when rounded down to the nearest multiple of 1000.

codec_level: Option<String>

Use Level (Mpeg2CodecLevel) to set the MPEG-2 level for the video output.

codec_profile: Option<String>

Use Profile (Mpeg2CodecProfile) to set the MPEG-2 profile for the video output.

dynamic_sub_gop: Option<String>

Choose Adaptive to improve subjective video quality for high-motion content. This will cause the service to use fewer B-frames (which infer information based on other frames) for high-motion portions of the video and more B-frames for low-motion portions. The maximum number of B-frames is limited by the value you provide for the setting B frames between reference frames (numberBFramesBetweenReferenceFrames).

framerate_control: Option<String>

If you are using the console, use the Framerate setting to specify the frame rate for this output. If you want to keep the same frame rate as the input video, choose Follow source. If you want to do frame rate conversion, choose a frame rate from the dropdown list or choose Custom. The framerates shown in the dropdown list are decimal approximations of fractions. If you choose Custom, specify your frame rate as a fraction. If you are creating your transcoding job sepecification as a JSON file without the console, use FramerateControl to specify which value the service uses for the frame rate for this output. Choose INITIALIZEFROMSOURCE if you want the service to use the frame rate from the input. Choose SPECIFIED if you want the service to use the frame rate you specify in the settings FramerateNumerator and FramerateDenominator.

framerate_conversion_algorithm: Option<String>

When set to INTERPOLATE, produces smoother motion during frame rate conversion.

framerate_denominator: Option<i64>

Frame rate denominator.

framerate_numerator: Option<i64>

Frame rate numerator - frame rate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

gop_closed_cadence: Option<i64>

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

gop_size: Option<f64>

GOP Length (keyframe interval) in frames or seconds. Must be greater than zero.

gop_size_units: Option<String>

Indicates if the GOP Size in MPEG2 is specified in frames or seconds. If seconds the system will convert the GOP Size into a frame count at run time.

hrd_buffer_initial_fill_percentage: Option<i64>

Percentage of the buffer that should initially be filled (HRD buffer model).

hrd_buffer_size: Option<i64>

Size of buffer (HRD buffer model) in bits. For example, enter five megabits as 5000000.

interlace_mode: Option<String>

Use Interlace mode (InterlaceMode) to choose the scan line type for the output. * Top Field First (TOPFIELD) and Bottom Field First (BOTTOMFIELD) produce interlaced output with the entire output having the same field polarity (top or bottom first). * Follow, Default Top (FOLLOWTOPFIELD) and Follow, Default Bottom (FOLLOWBOTTOMFIELD) use the same field polarity as the source. Therefore, behavior depends on the input scan type. - If the source is interlaced, the output will be interlaced with the same polarity as the source (it will follow the source). The output could therefore be a mix of "top field first" and "bottom field first". - If the source is progressive, the output will be interlaced with "top field first" or "bottom field first" polarity, depending on which of the Follow options you chose.

intra_dc_precision: Option<String>

Use Intra DC precision (Mpeg2IntraDcPrecision) to set quantization precision for intra-block DC coefficients. If you choose the value auto, the service will automatically select the precision based on the per-frame compression ratio.

max_bitrate: Option<i64>

Maximum bitrate in bits/second. For example, enter five megabits per second as 5000000.

min_i_interval: Option<i64>

Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. This setting is only used when Scene Change Detect is enabled. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

number_b_frames_between_reference_frames: Option<i64>

Number of B-frames between reference frames.

par_control: Option<String>

Using the API, enable ParFollowSource if you want the service to use the pixel aspect ratio from the input. Using the console, do this by choosing Follow source for Pixel aspect ratio.

par_denominator: Option<i64>

Pixel Aspect Ratio denominator.

par_numerator: Option<i64>

Pixel Aspect Ratio numerator.

quality_tuning_level: Option<String>

Use Quality tuning level (Mpeg2QualityTuningLevel) to specifiy whether to use single-pass or multipass video encoding.

rate_control_mode: Option<String>

Use Rate control mode (Mpeg2RateControlMode) to specifiy whether the bitrate is variable (vbr) or constant (cbr).

scene_change_detect: Option<String>

Scene change detection (inserts I-frames on scene changes).

slow_pal: Option<String>

Enables Slow PAL rate conversion. 23.976fps and 24fps input is relabeled as 25fps, and audio is sped up correspondingly.

softness: Option<i64>

Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image.

spatial_adaptive_quantization: Option<String>

Adjust quantization within each frame based on spatial variation of content complexity.

syntax: Option<String>

Produces a Type D-10 compatible bitstream (SMPTE 356M-2001).

telecine: Option<String>

Only use Telecine (Mpeg2Telecine) when you set Framerate (Framerate) to 29.970. Set Telecine (Mpeg2Telecine) to Hard (hard) to produce a 29.97i output from a 23.976 input. Set it to Soft (soft) to produce 23.976 output and leave converstion to the player.

temporal_adaptive_quantization: Option<String>

Adjust quantization within each frame based on temporal variation of content complexity.