Trait rusoto_mediaconvert::MediaConvert

pub trait MediaConvert {
    fn associate_certificate(
        &self, 
        input: AssociateCertificateRequest
    ) -> RusotoFuture<AssociateCertificateResponse, AssociateCertificateError>;

    fn cancel_job(
        &self, 
        input: CancelJobRequest
    ) -> RusotoFuture<CancelJobResponse, CancelJobError>;

    fn create_job(
        &self, 
        input: CreateJobRequest
    ) -> RusotoFuture<CreateJobResponse, CreateJobError>;

    fn create_job_template(
        &self, 
        input: CreateJobTemplateRequest
    ) -> RusotoFuture<CreateJobTemplateResponse, CreateJobTemplateError>;

    fn create_preset(
        &self, 
        input: CreatePresetRequest
    ) -> RusotoFuture<CreatePresetResponse, CreatePresetError>;

    fn create_queue(
        &self, 
        input: CreateQueueRequest
    ) -> RusotoFuture<CreateQueueResponse, CreateQueueError>;

    fn delete_job_template(
        &self, 
        input: DeleteJobTemplateRequest
    ) -> RusotoFuture<DeleteJobTemplateResponse, DeleteJobTemplateError>;

    fn delete_preset(
        &self, 
        input: DeletePresetRequest
    ) -> RusotoFuture<DeletePresetResponse, DeletePresetError>;

    fn delete_queue(
        &self, 
        input: DeleteQueueRequest
    ) -> RusotoFuture<DeleteQueueResponse, DeleteQueueError>;

    fn describe_endpoints(
        &self, 
        input: DescribeEndpointsRequest
    ) -> RusotoFuture<DescribeEndpointsResponse, DescribeEndpointsError>;

    fn disassociate_certificate(
        &self, 
        input: DisassociateCertificateRequest
    ) -> RusotoFuture<DisassociateCertificateResponse, DisassociateCertificateError>;

    fn get_job(
        &self, 
        input: GetJobRequest
    ) -> RusotoFuture<GetJobResponse, GetJobError>;

    fn get_job_template(
        &self, 
        input: GetJobTemplateRequest
    ) -> RusotoFuture<GetJobTemplateResponse, GetJobTemplateError>;

    fn get_preset(
        &self, 
        input: GetPresetRequest
    ) -> RusotoFuture<GetPresetResponse, GetPresetError>;

    fn get_queue(
        &self, 
        input: GetQueueRequest
    ) -> RusotoFuture<GetQueueResponse, GetQueueError>;

    fn list_job_templates(
        &self, 
        input: ListJobTemplatesRequest
    ) -> RusotoFuture<ListJobTemplatesResponse, ListJobTemplatesError>;

    fn list_jobs(
        &self, 
        input: ListJobsRequest
    ) -> RusotoFuture<ListJobsResponse, ListJobsError>;

    fn list_presets(
        &self, 
        input: ListPresetsRequest
    ) -> RusotoFuture<ListPresetsResponse, ListPresetsError>;

    fn list_queues(
        &self, 
        input: ListQueuesRequest
    ) -> RusotoFuture<ListQueuesResponse, ListQueuesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_job_template(
        &self, 
        input: UpdateJobTemplateRequest
    ) -> RusotoFuture<UpdateJobTemplateResponse, UpdateJobTemplateError>;

    fn update_preset(
        &self, 
        input: UpdatePresetRequest
    ) -> RusotoFuture<UpdatePresetResponse, UpdatePresetError>;

    fn update_queue(
        &self, 
        input: UpdateQueueRequest
    ) -> RusotoFuture<UpdateQueueResponse, UpdateQueueError>;
}

Trait representing the capabilities of the MediaConvert API. MediaConvert clients implement this trait.

Required methods

fn associate_certificate(
&self,
input: AssociateCertificateRequest
) -> RusotoFuture<AssociateCertificateResponse, AssociateCertificateError>

Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert.

fn cancel_job(
&self,
input: CancelJobRequest
) -> RusotoFuture<CancelJobResponse, CancelJobError>

Permanently cancel a job. Once you have canceled a job, you can't start it again.

fn create_job(
&self,
input: CreateJobRequest
) -> RusotoFuture<CreateJobResponse, CreateJobError>

Create a new transcoding job. For information about jobs and job settings, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

fn create_job_template(
&self,
input: CreateJobTemplateRequest
) -> RusotoFuture<CreateJobTemplateResponse, CreateJobTemplateError>

Create a new job template. For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

fn create_preset(
&self,
input: CreatePresetRequest
) -> RusotoFuture<CreatePresetResponse, CreatePresetError>

Create a new preset. For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

fn create_queue(
&self,
input: CreateQueueRequest
) -> RusotoFuture<CreateQueueResponse, CreateQueueError>

Create a new transcoding queue. For information about queues, see Working With Queues in the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html

fn delete_job_template(
&self,
input: DeleteJobTemplateRequest
) -> RusotoFuture<DeleteJobTemplateResponse, DeleteJobTemplateError>

Permanently delete a job template you have created.

fn delete_preset(
&self,
input: DeletePresetRequest
) -> RusotoFuture<DeletePresetResponse, DeletePresetError>

Permanently delete a preset you have created.

fn delete_queue(
&self,
input: DeleteQueueRequest
) -> RusotoFuture<DeleteQueueResponse, DeleteQueueError>

Permanently delete a queue you have created.

fn describe_endpoints(
&self,
input: DescribeEndpointsRequest
) -> RusotoFuture<DescribeEndpointsResponse, DescribeEndpointsError>

Send an request with an empty body to the regional API endpoint to get your account API endpoint.

fn disassociate_certificate(
&self,
input: DisassociateCertificateRequest
) -> RusotoFuture<DisassociateCertificateResponse, DisassociateCertificateError>

Removes an association between the Amazon Resource Name (ARN) of an AWS Certificate Manager (ACM) certificate and an AWS Elemental MediaConvert resource.

fn get_job(
&self,
input: GetJobRequest
) -> RusotoFuture<GetJobResponse, GetJobError>

Retrieve the JSON for a specific completed transcoding job.

fn get_job_template(
&self,
input: GetJobTemplateRequest
) -> RusotoFuture<GetJobTemplateResponse, GetJobTemplateError>

Retrieve the JSON for a specific job template.

fn get_preset(
&self,
input: GetPresetRequest
) -> RusotoFuture<GetPresetResponse, GetPresetError>

Retrieve the JSON for a specific preset.

fn get_queue(
&self,
input: GetQueueRequest
) -> RusotoFuture<GetQueueResponse, GetQueueError>

Retrieve the JSON for a specific queue.

fn list_job_templates(
&self,
input: ListJobTemplatesRequest
) -> RusotoFuture<ListJobTemplatesResponse, ListJobTemplatesError>

Retrieve a JSON array of up to twenty of your job templates. This will return the templates themselves, not just a list of them. To retrieve the next twenty templates, use the nextToken string returned with the array

fn list_jobs(
&self,
input: ListJobsRequest
) -> RusotoFuture<ListJobsResponse, ListJobsError>

Retrieve a JSON array of up to twenty of your most recently created jobs. This array includes in-process, completed, and errored jobs. This will return the jobs themselves, not just a list of the jobs. To retrieve the twenty next most recent jobs, use the nextToken string returned with the array.

fn list_presets(
&self,
input: ListPresetsRequest
) -> RusotoFuture<ListPresetsResponse, ListPresetsError>

Retrieve a JSON array of up to twenty of your presets. This will return the presets themselves, not just a list of them. To retrieve the next twenty presets, use the nextToken string returned with the array.

fn list_queues(
&self,
input: ListQueuesRequest
) -> RusotoFuture<ListQueuesResponse, ListQueuesError>

Retrieve a JSON array of up to twenty of your queues. This will return the queues themselves, not just a list of them. To retrieve the next twenty queues, use the nextToken string returned with the array.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Retrieve the tags for a MediaConvert resource.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Add tags to a MediaConvert queue, preset, or job template. For information about tagging, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/tagging-resources.html

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Remove tags from a MediaConvert queue, preset, or job template. For information about tagging, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/tagging-resources.html

fn update_job_template(
&self,
input: UpdateJobTemplateRequest
) -> RusotoFuture<UpdateJobTemplateResponse, UpdateJobTemplateError>

Modify one of your existing job templates.

fn update_preset(
&self,
input: UpdatePresetRequest
) -> RusotoFuture<UpdatePresetResponse, UpdatePresetError>

Modify one of your existing presets.

fn update_queue(
&self,
input: UpdateQueueRequest
) -> RusotoFuture<UpdateQueueResponse, UpdateQueueError>

Modify one of your existing queues.

Loading content...