Crate rusoto_mediaconvert

AWS Elemental MediaConvert

If you're using the service, you're probably looking for MediaConvertClient and MediaConvert.

Structs

AacSettings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AAC. The service accepts one of two mutually exclusive groups of AAC settings--VBR and CBR. To select one of these modes, set the value of Bitrate control mode (rateControlMode) to "VBR" or "CBR". In VBR mode, you control the audio quality with the setting VBR quality (vbrQuality). In CBR mode, you use the setting Bitrate (bitrate). Defaults and valid values depend on the rate control mode.

Ac3Settings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AC3.

AccelerationSettings
Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

AiffSettings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AIFF.

AncillarySourceSettings
Settings for ancillary captions source.

AssociateCertificateRequest

AssociateCertificateResponse

AudioCodecSettings
Audio codec settings (CodecSettings) under (AudioDescriptions) contains the group of settings related to audio encoding. The settings in this group vary depending on the value you choose for Audio codec (Codec). For each codec enum you choose, define the corresponding settings object. The following lists the codec enum, settings object pairs. * AAC, AacSettings * MP2, Mp2Settings * WAV, WavSettings * AIFF, AiffSettings * AC3, Ac3Settings * EAC3, Eac3Settings

AudioDescription
Description of audio output

AudioNormalizationSettings
Advanced audio normalization settings.

AudioSelector
Selector for Audio

AudioSelectorGroup
Group of Audio Selectors

AvailBlanking
Settings for Avail Blanking

BurninDestinationSettings
Burn-In Destination Settings.

CancelJobRequest

CancelJobResponse

CaptionDescription
Description of Caption output

CaptionDescriptionPreset
Caption Description for preset

CaptionDestinationSettings
Specific settings required by destination type. Note that burnindestinationsettings are not available if the source of the caption data is Embedded or Teletext.

CaptionSelector
Set up captions in your outputs by first selecting them from your input here.

CaptionSourceSettings
Source settings (SourceSettings) contains the group of settings for captions in the input.

ChannelMapping
Channel mapping (ChannelMapping) contains the group of fields that hold the remixing value for each channel. Units are in dB. Acceptable values are within the range from -60 (mute) through 6. A setting of 0 passes the input channel unchanged to the output channel (no attenuation or amplification).

CmafEncryptionSettings
Settings for CMAF encryption

CmafGroupSettings
Required when you set (Type) under (OutputGroups)>(OutputGroupSettings) to CMAFGROUPSETTINGS. Each output in a CMAF Output Group may only contain a single video, audio, or caption output.

ColorCorrector
Settings for color correction.

ContainerSettings
Container specific settings.

CreateJobRequest

CreateJobResponse

CreateJobTemplateRequest

CreateJobTemplateResponse

CreatePresetRequest

CreatePresetResponse

CreateQueueRequest

CreateQueueResponse

DashIsoEncryptionSettings
Specifies DRM settings for DASH outputs.

DashIsoGroupSettings
Required when you set (Type) under (OutputGroups)>(OutputGroupSettings) to DASHISOGROUP_SETTINGS.

Deinterlacer
Settings for deinterlacer

DeleteJobTemplateRequest

DeleteJobTemplateResponse

DeletePresetRequest

DeletePresetResponse

DeleteQueueRequest

DeleteQueueResponse

DescribeEndpointsRequest
DescribeEndpointsRequest

DescribeEndpointsResponse

DestinationSettings
Settings associated with the destination. Will vary based on the type of destination

DisassociateCertificateRequest

DisassociateCertificateResponse

DvbNitSettings
Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

DvbSdtSettings
Inserts DVB Service Description Table (NIT) at the specified table repetition interval.

DvbSubDestinationSettings
DVB-Sub Destination Settings

DvbSubSourceSettings
DVB Sub Source Settings

DvbTdtSettings
Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

Eac3Settings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3.

EmbeddedDestinationSettings
Settings specific to embedded/ancillary caption outputs, including 608/708 Channel destination number.

EmbeddedSourceSettings
Settings for embedded captions Source

Endpoint
Describes an account-specific API endpoint.

EsamManifestConfirmConditionNotification
ESAM ManifestConfirmConditionNotification defined by OC-SP-ESAM-API-I03-131025.

EsamSettings
Settings for Event Signaling And Messaging (ESAM). If you don't do ad insertion, you can ignore these settings.

EsamSignalProcessingNotification
ESAM SignalProcessingNotification data defined by OC-SP-ESAM-API-I03-131025.

F4vSettings
Settings for F4v container

FileGroupSettings
Required when you set (Type) under (OutputGroups)>(OutputGroupSettings) to FILEGROUPSETTINGS.

FileSourceSettings
Settings for File-based Captions in Source

FrameCaptureSettings
Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value FRAME_CAPTURE.

GetJobRequest

GetJobResponse

GetJobTemplateRequest

GetJobTemplateResponse

GetPresetRequest

GetPresetResponse

GetQueueRequest

GetQueueResponse

H264QvbrSettings
Settings for quality-defined variable bitrate encoding with the H.264 codec. Required when you set Rate control mode to QVBR. Not valid when you set Rate control mode to a value other than QVBR, or when you don't define Rate control mode.

H264Settings
Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value H_264.

H265QvbrSettings
Settings for quality-defined variable bitrate encoding with the H.265 codec. Required when you set Rate control mode to QVBR. Not valid when you set Rate control mode to a value other than QVBR, or when you don't define Rate control mode.

H265Settings
Settings for H265 codec

Hdr10Metadata
Use the "HDR master display information" (Hdr10Metadata) settings to correct HDR metadata or to provide missing metadata. These values vary depending on the input video and must be provided by a color grader. Range is 0 to 50,000; each increment represents 0.00002 in CIE1931 color coordinate. Note that these settings are not color correction. Note that if you are creating HDR outputs inside of an HLS CMAF package, to comply with the Apple specification, you must use the following settings. Set "MP4 packaging type" (writeMp4PackagingType) to HVC1 (HVC1). Set "Profile" (H265Settings > codecProfile) to Main10/High (MAIN10HIGH). Set "Level" (H265Settings > codecLevel) to 5 (LEVEL5).

HlsCaptionLanguageMapping
Caption Language Mapping

HlsEncryptionSettings
Settings for HLS encryption

HlsGroupSettings
Required when you set (Type) under (OutputGroups)>(OutputGroupSettings) to HLSGROUPSETTINGS.

HlsSettings
Settings for HLS output groups

Id3Insertion
To insert ID3 tags in your output, specify two values. Use ID3 tag (Id3) to specify the base 64 encoded string and use Timecode (TimeCode) to specify the time when the tag should be inserted. To insert multiple ID3 tags in your output, create multiple instances of ID3 insertion (Id3Insertion).

ImageInserter
Enable the image inserter feature to include a graphic overlay on your video. Enable or disable this feature for each input or output individually. This setting is disabled by default.

Input
Specifies media input

InputClipping
To transcode only portions of your input (clips), include one Input clipping (one instance of InputClipping in the JSON job file) for each input clip. All input clips you specify will be included in every output of the job.

InputDecryptionSettings
Settings for decrypting any input files that you encrypt before you upload them to Amazon S3. MediaConvert can decrypt files only when you use AWS Key Management Service (KMS) to encrypt the data key that you use to encrypt your content.

InputTemplate
Specified video input in a template.

InsertableImage
Settings that specify how your still graphic overlay appears.

Job
Each job converts an input file into an output file or files. For more information, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

JobSettings
JobSettings contains all the transcode settings for a job.

JobTemplate
A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

JobTemplateSettings
JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

ListJobTemplatesRequest

ListJobTemplatesResponse

ListJobsRequest

ListJobsResponse

ListPresetsRequest

ListPresetsResponse

ListQueuesRequest

ListQueuesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

M2tsSettings
MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

M2tsScte35Esam
Settings for SCTE-35 signals from ESAM. Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

M3u8Settings
Settings for TS segments in HLS

MediaConvertClient
A client for the MediaConvert API.

MotionImageInserter
Overlay motion graphics on top of your video at the time that you specify.

MotionImageInsertionFramerate
For motion overlays that don't have a built-in frame rate, specify the frame rate of the overlay in frames per second, as a fraction. For example, specify 24 fps as 24/1. The overlay frame rate doesn't need to match the frame rate of the underlying video.

MotionImageInsertionOffset
Specify the offset between the upper-left corner of the video frame and the top left corner of the overlay.

MovSettings
Settings for MOV Container.

Mp2Settings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value MP2.

Mp4Settings
Settings for MP4 Container

Mpeg2Settings
Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value MPEG2.

MsSmoothEncryptionSettings
If you are using DRM, set DRM System (MsSmoothEncryptionSettings) to specify the value SpekeKeyProvider.

MsSmoothGroupSettings
Required when you set (Type) under (OutputGroups)>(OutputGroupSettings) to MSSMOOTHGROUP_SETTINGS.

NielsenConfiguration
Settings for Nielsen Configuration

NoiseReducer
Enable the Noise reducer (NoiseReducer) feature to remove noise from your video output if necessary. Enable or disable this feature for each output individually. This setting is disabled by default. When you enable Noise reducer (NoiseReducer), you must also select a value for Noise reducer filter (NoiseReducerFilter).

NoiseReducerFilterSettings
Settings for a noise reducer filter

NoiseReducerSpatialFilterSettings
Noise reducer filter settings for spatial filter.

Output
An output object describes the settings for a single output file or stream in an output group.

OutputChannelMapping
OutputChannel mapping settings.

OutputDetail
Details regarding output

OutputGroup
Group of outputs

OutputGroupDetail
Contains details about the output groups specified in the job settings.

OutputGroupSettings
Output Group settings, including type

OutputSettings
Specific settings for this type of output.

Preset
A preset is a collection of preconfigured media conversion settings that you want MediaConvert to apply to the output during the conversion process.

PresetSettings
Settings for preset

ProresSettings
Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value PRORES.

Queue
You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

Rectangle
Use Rectangle to identify a specific area of the video frame.

RemixSettings
Use Manual audio remixing (RemixSettings) to adjust audio levels for each audio channel in each output of your job. With audio remixing, you can output more or fewer audio channels than your input audio source provides.

ReservationPlan
Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

ReservationPlanSettings
Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

ResourceTags
The Amazon Resource Name (ARN) and tags for an AWS Elemental MediaConvert resource.

S3DestinationSettings
Settings associated with S3 destination

S3EncryptionSettings
Settings for how your job outputs are encrypted as they are uploaded to Amazon S3.

SccDestinationSettings
Settings for SCC caption output.

SpekeKeyProvider
Settings for use with a SPEKE key provider

StaticKeyProvider
Use these settings to set up encryption with a static key provider.

TagResourceRequest

TagResourceResponse

TeletextDestinationSettings
Settings for Teletext caption output

TeletextSourceSettings
Settings specific to Teletext caption sources, including Page number.

TimecodeBurnin
Timecode burn-in (TimecodeBurnIn)--Burns the output timecode and specified prefix into the output.

TimecodeConfig
These settings control how the service handles timecodes throughout the job. These settings don't affect input clipping.

TimedMetadataInsertion
Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in your job. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

Timing
Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

TrackSourceSettings
Settings specific to caption sources that are specfied by track number. Sources include IMSC in IMF.

TtmlDestinationSettings
Settings specific to TTML caption outputs, including Pass style information (TtmlStylePassthrough).

UntagResourceRequest

UntagResourceResponse

UpdateJobTemplateRequest

UpdateJobTemplateResponse

UpdatePresetRequest

UpdatePresetResponse

UpdateQueueRequest

UpdateQueueResponse

VideoCodecSettings
Video codec settings, (CodecSettings) under (VideoDescription), contains the group of settings related to video encoding. The settings in this group vary depending on the value you choose for Video codec (Codec). For each codec enum you choose, define the corresponding settings object. The following lists the codec enum, settings object pairs. * H264, H264Settings * H265, H265Settings * MPEG2, Mpeg2Settings * PRORES, ProresSettings * FRAME_CAPTURE, FrameCaptureSettings

VideoDescription
Settings for video outputs

VideoDetail
Contains details about the output's video stream

VideoPreprocessor
Find additional transcoding features under Preprocessors (VideoPreprocessors). Enable the features at each output individually. These features are disabled by default.

VideoSelector
Selector for video.

WavSettings
Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value WAV.

Enums

AssociateCertificateError
Errors returned by AssociateCertificate

CancelJobError
Errors returned by CancelJob

CreateJobError
Errors returned by CreateJob

CreateJobTemplateError
Errors returned by CreateJobTemplate

CreatePresetError
Errors returned by CreatePreset

CreateQueueError
Errors returned by CreateQueue

DeleteJobTemplateError
Errors returned by DeleteJobTemplate

DeletePresetError
Errors returned by DeletePreset

DeleteQueueError
Errors returned by DeleteQueue

DescribeEndpointsError
Errors returned by DescribeEndpoints

DisassociateCertificateError
Errors returned by DisassociateCertificate

GetJobError
Errors returned by GetJob

GetJobTemplateError
Errors returned by GetJobTemplate

GetPresetError
Errors returned by GetPreset

GetQueueError
Errors returned by GetQueue

ListJobTemplatesError
Errors returned by ListJobTemplates

ListJobsError
Errors returned by ListJobs

ListPresetsError
Errors returned by ListPresets

ListQueuesError
Errors returned by ListQueues

ListTagsForResourceError
Errors returned by ListTagsForResource

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateJobTemplateError
Errors returned by UpdateJobTemplate

UpdatePresetError
Errors returned by UpdatePreset

UpdateQueueError
Errors returned by UpdateQueue

Traits

MediaConvert
Trait representing the capabilities of the MediaConvert API. MediaConvert clients implement this trait.